requirements.txt
hopsworks[python,great-expectations]
streamlit==1.28.2
email-validator==2.2.0
pydantic-settings>=2.6.1
geopy==2.4.1
openmeteo-requests
requests-cache==1.2.0
retry-requests==2.0.0
xgboost==2.0.3
scikit-learn==1.2.2
matplotlib==3.8.3
plotly
seaborn
nbformat
Faker
invoke
python-dotenv
#feldera==0.41
